# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: RestMail API
  version: 1.0.0
  description: >
    RestMail is a modern email platform with REST API access, supporting traditional
    SMTP/IMAP/POP3 gateways and the RESTMAIL server-to-server protocol. This spec
    covers all user-facing, admin, internal gateway, and RESTMAIL endpoints.
  license:
    name: MIT
servers:
  - url: /
    description: Default server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/admin/bans:
    get:
      operationId: listBans
      tags: ["Admin: Bans"]
      summary: List IP bans
      parameters:
        - name: protocol
          in: query
          schema:
            type: string
            enum: [smtp, imap, pop3, all]
        - name: ip
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Ban list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Ban"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
    post:
      operationId: createBan
      tags: ["Admin: Bans"]
      summary: Create an IP ban
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBanRequest"
      responses:
        "201":
          description: Ban created
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/Ban"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/admin/bans/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    delete:
      operationId: deleteBan
      tags: ["Admin: Bans"]
      summary: Delete a ban by ID
      responses:
        "204":
          description: Ban deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/bans/ip/{ip}:
    parameters:
      - name: ip
        in: path
        required: true
        schema:
          type: string
        description: IP address to unban
    delete:
      operationId: unbanIP
      tags: ["Admin: Bans"]
      summary: Remove all bans for a specific IP
      responses:
        "204":
          description: IP unbanned
        "404":
          $ref: "#/components/responses/NotFound"

  # ═══════════════════════════════════════════════════════════════
  # Admin: Logs
  # ═══════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Resource ID
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: not_found
            message:
              type: string
              example: Resource not found
            details:
              type: object
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: validation_failed
            message:
              type: string
              example: Request body failed validation
            details:
              type: object
              properties:
                fields:
                  type: object
                  additionalProperties:
                    type: string
    Pagination:
      type: object
      properties:
        cursor:
          type: string
          description: Opaque cursor for the next page
        has_more:
          type: boolean
        total:
          type: integer
          format: int64

    # ─────────────────────────────────────────────────────────────
    # Health
    # ─────────────────────────────────────────────────────────────
...
    Ban:
      type: object
      required: [id, ip, reason, protocol, created_by, created_at]
      properties:
        id:
          type: integer
        ip:
          type: string
        reason:
          type: string
        protocol:
          type: string
          enum: [smtp, imap, pop3, all]
        created_by:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateBanRequest:
      type: object
      required: [ip, protocol, created_by]
      properties:
        ip:
          type: string
        reason:
          type: string
        protocol:
          type: string
          enum: [smtp, imap, pop3, all]
        duration:
          type: string
          description: Duration string (e.g. "24h", "7d"), omit for permanent
        created_by:
          type: string

    # ─────────────────────────────────────────────────────────────
    # Activity Log
    # ─────────────────────────────────────────────────────────────
...